Move several functions from server.js file to controllers.js
